+++ 3dEngine/src/scene/ui/widget/staticbitmap/StaticBitmap.cpp
@@ -40,7 +40,7 @@ namespace urchin {
     }
 
     void StaticBitmap::prepareWidgetRendering(float) {
-        updateTranslateVector(bitmapRenderer.get(), Vector2<float>(getGlobalPositionX(), getGlobalPositionY()));
+        updateTranslateVector(bitmapRenderer.get(), Vector2<int>(getGlobalPositionX(), getGlobalPositionY()));
         bitmapRenderer->enableRenderer();
     }
